consistent use of mock reward model
