+++ prompting/validators/mock.py
@@ -45,7 +45,23 @@ class MockGatingModel(BaseGatingModel):
         pass
 
 
-class MockRewardModel(torch.nn.Module):
+class MockRewardModel(BaseRewardModel):
+    @property
+    def name(self) -> str:
+        return self.mock_name
+
+    def __init__(self, mock_name: str = "MockReward"):
+        super().__init__()
+        self.mock_name = mock_name
+        self.question_blacklist = {}
+
+    def apply(self, prompt: str, completion: List[str], name: str) -> torch.FloatTensor:
+        mock_reward = torch.tensor([1 for _ in completion], dtype=torch.float32)
+        return mock_reward, mock_reward
+
+    def reset(self):
+        return self
+
     def reward(
         self,
         completions_with_prompt: List[str],
