+++ prompting/validators/reward/__init__.py
@@ -6,7 +6,6 @@ from .open_assistant import OpenAssistantRewardModel
 from .reciprocate import ReciprocateRewardModel
 from .relevance import RelevanceRewardModel
 from .reward import BaseRewardModel
-from .reward import MockRewardModel
 from .dahoas import DahoasRewardModel
 from .diversity import DiversityRewardModel
 from .prompt import PromptRewardModel
