+++ prompting/validators/utils.py
@@ -23,7 +23,7 @@ import copy
 import bittensor as bt
 import prompting.validators as validators
 from prompting.validators.misc import ttl_get_block
-from prompting.validators.reward import MockRewardModel
+from prompting.validators.mock import MockRewardModel
 
 
 def should_reinit_wandb(self):
